Rename fn_args to args
