+++ src/metrics/mod.rs
@@ -1,8 +1,8 @@
 pub mod cognitive;
 pub mod cyclomatic;
 pub mod exit;
-pub mod fn_args;
 pub mod halstead;
 pub mod loc;
 pub mod mi;
+pub mod nargs;
 pub mod nom;
