+++ src/traits.rs
@@ -7,12 +7,12 @@ use crate::checker::Checker;
 use crate::cognitive::Cognitive;
 use crate::cyclomatic::Cyclomatic;
 use crate::exit::Exit;
-use crate::fn_args::NArgs;
 use crate::getter::Getter;
 use crate::halstead::Halstead;
 use crate::langs::*;
 use crate::loc::Loc;
 use crate::mi::Mi;
+use crate::nargs::NArgs;
 use crate::node::Node;
 use crate::nom::Nom;
 use crate::parser::Filter;
